Moved ModificatorInfo and kind, into ModificatorTypeMapper
