--- national-connector/cda-generator/src/main/java/dk/sundhedsdatastyrelsen/ncpeh/cda/model/ModificatorKind.java
@@ -1,9 +1,0 @@
-package dk.sundhedsdatastyrelsen.ncpeh.cda.model;
-
-public enum ModificatorKind {
-    HEALTHCARE_PROFESSIONAL,
-    PARTLY_DEFINED_EFFECTUATOR,
-    ORGANISATION,
-    OTHER,
-    UNKNOWN
-}
